requirements.txt
# Core dependencies
z3-solver>=4.12.0
python-dotenv>=1.0.0

# LLM and Azure OpenAI dependencies
openai>=1.0.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20

# Visualization dependencies
graphviz>=0.20.0

# Standard library dependencies (included with Python)
# subprocess, base64, os, re, ast, html are built-in modules 